Add css elements, home button and make minor edits
Add multiple css elements, such as images sizes -which were moved
from the html file to the css one-; font sizes, weight and families
for some specifics pieces of text; background color for the whole web,
and also different elements for the new home button, which replaces the
old home link.

Moreover, edit the type of some elements placed on lists, which are now
type p in order to generate a spacing between the elements of the lists.

Finally, correct a spelling mistake in the readme.