Add buttons on all pages - make aesthetic changes
Add buttons on all the pages of the website that redirects to all the
other pages in order to easily access any page without having to go to
home and then to the page.

Also, make some aesthetic changes such as the addition horizontal
lines that separate the buttons from the rest of the body, besides a
font change to the creator signature.