Edit some types and classes used on html files
Edit some of the types and classes used on the html files, removing the
un-list class from the index's list since it's not necessary having the
type p, which has the same properties applied.

Also, change the use of type p on the unordered lists of the recipes
pages to type div with the class un-list applied, this because type p
implies a spacing between every line, which is not desired for the
elements of those lists.

Moreover, change the home-button id to a class, since it is used
multiple times on the web, and also correct some indentation problems
on html files.