Create sidebar and remove links on top and bottom
Change the structure of the web -only on index- by creating a sidebar
that contains links which are moved from the links sections on top and
bottom of the page to there. The reason of this change is the web's
aesthetics.

Also, move the main content of the page to the right of this sidebar and
remove the link sections on top and bottom of the page.