Revert last change to improve mob visualization
Revert the changes of the last commit by removing the set viewport,
since it isn't helpful to improve mobile visualization.

setting a viewport with the width
of the determined device where the web is being browsed on. Also, add
a horizontal scrollbar to make possible surfing through the whole
content of the web.